modify lend and customer

+++ interal1-bcml-service/src/main/java/com/koukio/repository/LendRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.*;
 import com.koukio.entity.Customer;
 import com.koukio.entity.Lend;
 
-public interface LendRepository extends Repository<Lend, Long> {
+public interface LendRepository extends Repository<Lend, Integer> {
 
 	 Page<Lend> findAll(Pageable pageable);
